+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Changed
+- `separate_apps` example now uses the negotiation protocol provided by comit-sdk.
+
 ## [0.4.0] - 2019-11-20
 
 ### Fixed